add LoadingComponent for reusability, and fix minor UI issues
