+++ src/people-container/components/PersonForm.js
@@ -3,15 +3,15 @@
 import React, {Component} from 'react';
 import {observable} from 'mobx';
 import Modal from 'react-modal';
-import {PulseLoader} from 'react-spinners';
 import Autocomplete from 'react-autocomplete';
+import {PropTypes} from 'prop-types';
+import {observer} from 'mobx-react';
+
 import css from './PersonForm.css';
 import closeIcon from '../../assets/icon_close.svg';
 import addIcon from '../../assets/icon_add_b.svg';
-import {PropTypes} from 'prop-types';
-import {observer} from 'mobx-react';
 import skillsStore from '../skills-store';
-import utilityStore from '../../utils/utility-store';
+import Loading from '../../loading-component/LoadingComponent';
 
 @observer
 class PersonForm extends Component {
@@ -367,9 +367,7 @@ class PersonForm extends Component {
             </div>
           </form>
         </div>
-        <Modal isOpen={utilityStore.isWaiting} style={waitingModalStyle}>
-          Please wait<PulseLoader color={'white'} />
-        </Modal>
+        <Loading />
       </Modal>
     );
   }
@@ -390,23 +388,6 @@ const modalStyle = {
     maxHeight: '1000px',
     maxWidth: '1220px',
     margin: '100px auto auto',
-  },
-};
-
-const waitingModalStyle = {
-  overlay: {
-    backgroundColor: 'rgba(0, 0, 0, 0.6)',
-  },
-  content: {
-    background: 'transparent',
-    color: 'white',
-    border: '0',
-    fontSize: '34px',
-    textAlign: 'center',
-    padding: '20px 30px',
-    maxWidth: '300px',
-    maxHeight: '200px',
-    margin: '150px auto auto',
   },
 };
